# openapi/agent-context-v1.json
{
  "components": {
    "schemas": {
      "AcceptedEventV1": {
        "additionalProperties": false,
        "description": "Result for an event that was newly stored or already known.",
        "properties": {
          "event_id": {
            "format": "uuid7",
            "title": "Event Id",
            "type": "string"
          },
          "status": {
            "enum": [
              "accepted",
              "existing"
            ],
            "title": "Status",
            "type": "string"
          },
          "stream_sequence": {
            "exclusiveMinimum": 0.0,
            "title": "Stream Sequence",
            "type": "integer"
          }
        },
        "required": [
          "event_id",
          "status",
          "stream_sequence"
        ],
        "title": "AcceptedEventV1",
        "type": "object"
      },
      "ContentClaimV1": {
        "additionalProperties": false,
        "description": "Producer-declared identity and metadata for sanitized content.",
        "properties": {
          "content_id": {
            "title": "Content Id",
            "type": "string"
          },
          "content_sha256": {
            "pattern": "^[0-9a-f]{64}$",
            "title": "Content Sha256",
            "type": "string"
          },
          "media_type": {
            "title": "Media Type",
            "type": "string"
          },
          "uncompressed_bytes": {
            "minimum": 0,
            "title": "Uncompressed Bytes",
            "type": "integer"
          }
        },
        "required": [
          "content_id",
          "content_sha256",
          "media_type",
          "uncompressed_bytes"
        ],
        "title": "ContentClaimV1",
        "type": "object"
      },
      "ContentDisposition": {
        "description": "Outcome of the redaction and content-retention boundary.",
        "enum": [
          "sanitized",
          "metadata_only",
          "dropped_redaction_failure",
          "purged"
        ],
        "title": "ContentDisposition",
        "type": "string"
      },
      "ErrorDetailV1": {
        "description": "Fixed, content-free error description; ``code`` is stable, ``message`` is not.",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetailV1",
        "type": "object"
      },
      "ErrorResponseV1": {
        "description": "Body of every failure the route answers before or instead of a batch outcome.",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetailV1"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "error",
          "request_id"
        ],
        "title": "ErrorResponseV1",
        "type": "object"
      },
...
      "EventDraftV1": {
        "additionalProperties": false,
        "description": "Producer-created event awaiting server content resolution and sealing.",
        "properties": {
          "content_claims": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/ContentClaimV1"
            },
            "title": "Content Claims",
            "type": "array"
          },
          "context": {
            "$ref": "#/components/schemas/EventContextV1"
          },
          "event_id": {
            "format": "uuid7",
            "title": "Event Id",
            "type": "string"
          },
          "event_type": {
            "title": "Event Type",
            "type": "string"
          },
          "idempotency_key": {
            "title": "Idempotency Key",
            "type": "string"
          },
          "observed_at": {
            "format": "date-time",
            "pattern": "(?:[zZ]|[+-]00:00)$",
            "title": "Observed At",
            "type": "string"
          },
          "occurred_at": {
            "format": "date-time",
            "pattern": "(?:[zZ]|[+-]00:00)$",
            "title": "Occurred At",
            "type": "string"
          },
          "payload": {
            "additionalProperties": {
              "$ref": "#/components/schemas/FrozenJsonValue"
            },
            "title": "Payload",
            "type": "object"
          },
          "producer": {
            "$ref": "#/components/schemas/ProducerV1"
          },
          "redaction": {
            "$ref": "#/components/schemas/EventRedactionSummaryV1"
          },
          "schema_version": {
            "const": "1.0.0",
            "default": "1.0.0",
            "title": "Schema Version",
            "type": "string"
          },
          "stream_id": {
            "title": "Stream Id",
            "type": "string"
          },
          "trace": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TraceContextV1"
              },
              {
                "type": "null"
              }
            ],
            "default": null
          }
        },
        "required": [
          "event_type",
          "stream_id",
          "occurred_at",
          "observed_at",
          "producer",
          "payload",
          "redaction",
          "idempotency_key"
        ],
        "title": "EventDraftV1",
        "type": "object"
      },
...
      "FrozenJsonValue": {
        "anyOf": [
          {
            "type": "boolean"
          },
          {
            "type": "integer"
          },
          {
            "type": "number"
          },
          {
            "type": "string"
          },
          {
            "items": {
              "$ref": "#/components/schemas/FrozenJsonValue"
            },
            "type": "array"
          },
          {
            "additionalProperties": {
              "$ref": "#/components/schemas/FrozenJsonValue"
            },
            "type": "object"
          },
          {
            "type": "null"
          }
        ]
      },
      "IngestBatchRequestV1": {
        "additionalProperties": false,
        "description": "A bounded, atomically processed collection of event drafts and bytes.",
        "properties": {
          "batch_id": {
            "format": "uuid7",
            "title": "Batch Id",
            "type": "string"
          },
          "content_items": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/SanitizedContentItemV1"
            },
            "title": "Content Items",
            "type": "array"
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/EventDraftV1"
            },
            "maxItems": 500,
            "minItems": 1,
            "title": "Events",
            "type": "array"
          }
        },
        "required": [
          "batch_id",
          "events"
        ],
        "title": "IngestBatchRequestV1",
        "type": "object"
      },
      "IngestBatchResponseV1": {
        "additionalProperties": false,
        "description": "Atomic ingestion outcome for a submitted batch.",
        "properties": {
          "accepted": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/AcceptedEventV1"
            },
            "title": "Accepted",
            "type": "array"
          },
          "batch_id": {
            "format": "uuid7",
            "title": "Batch Id",
            "type": "string"
          },
          "rejected": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/RejectedEventV1"
            },
            "title": "Rejected",
            "type": "array"
          }
        },
        "required": [
          "batch_id"
        ],
        "title": "IngestBatchResponseV1",
        "type": "object"
      },
...
      "ProducerV1": {
        "additionalProperties": false,
        "description": "Identity of the adapter that observed an event.",
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          },
          "producer_id": {
            "title": "Producer Id",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "producer_id",
          "name",
          "version"
        ],
        "title": "ProducerV1",
        "type": "object"
      },
...
      "RedactionReportV1": {
        "additionalProperties": false,
        "description": "Secret-safe aggregate report for one sanitized or discarded content item.",
        "properties": {
          "disposition": {
            "$ref": "#/components/schemas/ContentDisposition"
          },
          "error_class": {
            "anyOf": [
              {
                "pattern": "^[A-Za-z][A-Za-z0-9_.]{0,127}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Error Class"
          },
          "failed_detector": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Failed Detector"
          },
          "failed_detector_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Failed Detector Version"
          },
          "findings": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/RedactionFindingSummaryV1"
            },
            "title": "Findings",
            "type": "array"
          },
          "metadata_only_reason": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MetadataOnlyReason"
              },
              {
                "type": "null"
              }
            ],
            "default": null
          },
          "policy_version": {
            "title": "Policy Version",
            "type": "string"
          }
        },
        "required": [
          "policy_version",
          "disposition"
        ],
        "title": "RedactionReportV1",
        "type": "object"
      },
      "RejectedEventV1": {
        "additionalProperties": false,
        "description": "Public, safe result for a recoverable event-specific ingestion rejection.",
        "properties": {
          "error_code": {
            "pattern": "^[a-z][a-z0-9_]{0,127}$",
            "title": "Error Code",
            "type": "string"
          },
          "event_id": {
            "format": "uuid7",
            "title": "Event Id",
            "type": "string"
          },
          "retryable": {
            "title": "Retryable",
            "type": "boolean"
          }
        },
        "required": [
          "event_id",
          "error_code",
          "retryable"
        ],
        "title": "RejectedEventV1",
        "type": "object"
      },
      "SanitizedContentItemV1": {
        "additionalProperties": false,
        "description": "One verified, sanitized content payload supplied with an ingestion batch.",
        "properties": {
          "claim": {
            "$ref": "#/components/schemas/ContentClaimV1"
          },
          "redaction_report": {
            "$ref": "#/components/schemas/RedactionReportV1"
          },
          "sanitized_bytes_base64": {
            "title": "Sanitized Bytes Base64",
            "type": "string"
          }
        },
        "required": [
          "claim",
          "sanitized_bytes_base64",
          "redaction_report"
        ],
        "title": "SanitizedContentItemV1",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Agent Context Ingestion API",
    "version": "v1"
  },
  "openapi": "3.1.0",
  "paths": {
    "/v1/ingestion/batches": {
      "post": {
        "operationId": "ingest_batch_v1_ingestion_batches_post",
        "parameters": [
          {
            "description": "Must equal the batch_id in the body.",
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IngestBatchRequestV1"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestBatchResponseV1"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseV1"
                }
              }
            },
            "description": "Idempotency-Key header absent: `idempotency_key_required`.",
            "headers": {
              "x-request-id": {
                "description": "Request ID: echoed when a short safe token, otherwise generated.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseV1"
                }
              }
            },
            "description": "Missing or invalid producer credential: `missing_credential`, `invalid_credential`.",
            "headers": {
              "WWW-Authenticate": {
                "description": "Always `Bearer`.",
                "schema": {
                  "type": "string"
                }
              },
              "x-request-id": {
                "description": "Request ID: echoed when a short safe token, otherwise generated.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseV1"
                }
              }
            },
            "description": "Credential lacks the scope or names another producer: `insufficient_scope`, `producer_mismatch`.",
            "headers": {
              "x-request-id": {
                "description": "Request ID: echoed when a short safe token, otherwise generated.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestBatchResponseV1"
                }
              }
            },
            "description": "Whole batch rejected, nothing stored: `idempotency_conflict`, `stream_quarantined`; other events are `batch_rejected`.",
            "headers": {
              "x-request-id": {
                "description": "Request ID: echoed when a short safe token, otherwise generated.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseV1"
                }
              }
            },
            "description": "Request body exceeds the size limit: `request_too_large`.",
            "headers": {
              "x-request-id": {
                "description": "Request ID: echoed when a short safe token, otherwise generated.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/IngestBatchResponseV1"
                    },
                    {
                      "$ref": "#/components/schemas/ErrorResponseV1"
                    }
                  ],
                  "title": "Response 422 Ingest Batch V1 Ingestion Batches Post"
                }
              }
            },
            "description": "Error envelope: `invalid_request_schema`, `idempotency_key_mismatch`. Batch response, nothing stored: `duplicate_idempotency_key`, `duplicate_content_id`, `unsupported_media_type`, `invalid_content_encoding`, `content_requires_redaction`, `content_resolution_mismatch`.",
            "headers": {
              "x-request-id": {
                "description": "Request ID: echoed when a short safe token, otherwise generated.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseV1"
                }
              }
            },
            "description": "Unexpected failure: `internal_error`.",
            "headers": {
              "x-request-id": {
                "description": "Request ID: echoed when a short safe token, otherwise generated.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/IngestBatchResponseV1"
                    },
                    {
                      "$ref": "#/components/schemas/ErrorResponseV1"
                    }
                  ],
                  "title": "Response 503 Ingest Batch V1 Ingestion Batches Post"
                }
              }
            },
            "description": "Error envelope: `ingestion_unavailable`, `auth_overloaded`, `service_unavailable`. Batch response (retryable outage): rejected events carry `service_unavailable` with `retryable` true.",
            "headers": {
              "Retry-After": {
                "description": "Seconds to wait before retrying.",
                "schema": {
                  "type": "integer"
                }
              },
              "x-request-id": {
                "description": "Request ID: echoed when a short safe token, otherwise generated.",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "summary": "Ingest Batch",
        "tags": [
          "ingestion"
        ]
      }
    }
  }
}